#code-optimization : count() 함수 최적화
기존 '@Query'를 사용해서 값을 가져오는 것에서 'countByPost()'를 long 타입으로 받는 jpa 메소드를 사용하여 코드 최적화

+++ src/main/java/com/project/sns/service/PostService.java
@@ -85,7 +85,7 @@ public class PostService {
         alarmEntityRepository.save(AlarmEntity.of(postEntity.getUser(), AlarmType.NEW_LIKE_ON_POST, new AlarmArgs(userEntity.getId(), postEntity.getId())));
     }
 
-    public int likeCount(Integer postId) {
+    public long likeCount(Integer postId) {
         PostEntity postEntity = getPostEntityOrException(postId);
 
         // count like
